完善模板生成

+++ src/DAFigure/DAChartYDataPicker.cpp
@@ -292,9 +292,9 @@ QLineF DAChartYDataPicker::getCurveLine(QwtPlotCurve* curve, double x) const
                 return (x < pos.x());
             });
 
-            if (index == -1 && x == curve->sample(curve->dataSize() - 1).x()) {
+            if (index == -1 && x == curve->sample(static_cast<int>(curve->dataSize() - 1)).x()) {
                 // the last sample is excluded from qwtUpperSampleIndex
-                index = curve->dataSize() - 1;
+                index = static_cast<int>(curve->dataSize() - 1);
             }
 
             if (index > 0) {
@@ -315,9 +315,9 @@ double DAChartYDataPicker::getBarValue(QwtPlotBarChart* bar, double x) const
             int index = qwtUpperSampleIndex< QPointF >(*bar->data(), x, [](const double& x1, const QPointF& p) -> bool {
                 return (x1 < p.x());
             });
-            if (index == -1 && x == bar->sample(bar->dataSize() - 1).x()) {
+            if (index == -1 && x == bar->sample(static_cast<int>(bar->dataSize() - 1)).x()) {
                 // the last sample is excluded from qwtUpperSampleIndex
-                index = bar->dataSize() - 1;
+                index = static_cast<int>(bar->dataSize() - 1);
             }
             if (index > 0) {
                 return bar->sample(index).y();
